# POC PGO configuration